# api.yml
openapi: "3.0.2"
info:
  title: Tournament manager
  version: "1.0"
servers:
  - url: http://localhost:8080/api/v1
    description: localhost
  - url: https://dev.juno.nponsard.net/api/v1
    description: dev server
  - url: https://tournament.juno.nponsard.net/api/v1
    description: prod server
paths:
...
  /tournaments:
    get:
      summary: Get all tournaments
      tags:
        - "tournaments"

      parameters:
        - in: query
          name: limit
          schema:
            type: number
          description: Number of matches to return
          required: false
        - in: query
          name: offset
          schema:
            type: number
          description: Offset of the matches to return, default 0
          required: false
        - in: query
          name: search
          schema:
            type: string
          description: Search for string in tournaments name
          required: false

      responses:
        "200":
          description: Tournaments found
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournaments:
                    type: array
                    items:
                      $ref: "#/components/schemas/tournament"
                  total:
                    type: number
                    description: Total number of tournaments
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    post:
      summary: Create a tournament
      description: Create a tournament, the user must be an admin or organizer of the tournament
      tags:
        - "tournaments"
      security:
        - jwt: []
      requestBody:
        description: Tournament info
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: number
                name:
                  type: string
                start_date:
                  type: string
                end_date:
                  type: string
                description:
                  type: string
                max_teams:
                  type: number
                game_name:
                  type: string

      responses:
        "200":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tournament"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
  /tournaments/{id}:
    get:
      summary: Get a tournament
      tags:
        - "tournaments"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Tournament found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tournament"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    delete:
      summary: Delete a tournament
      tags:
        - "tournaments"
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Tournament deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    patch:
      summary: Update a tournament
      description: Update a tournament, the user must be admin or organizer of the tournament, can’t update type because it requires too much changes.
      tags:
        - "tournaments"
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        description: Tournament info
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  nullable: true
                start_date:
                  type: string
                  nullable: true
                end_date:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
                max_teams:
                  type: number
                  nullable: true
                game_name:
                  type: string
                  nullable: true
                status:
                  type: number
                  nullable: true
      responses:
        "200":
          description: Tournament updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tournament"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: JWT token
  responses:
    "400BadRequest":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
    "401Unauthorized":
      description: Invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
    "403Forbidden":
      description: not enough permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
    "500InternalServerError":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
  schemas:
    message:
      type: object
      description: object containing a message
      properties:
        message:
          type: string
...
    tournament:
      type: object
      description: Tournament object
      properties:
        id:
          type: number
        type:
          type: number
        name:
          type: string
        start_date:
          type: number
        end_date:
          type: number
        description:
          type: string
        status:
          type: number
        game_name:
          type: string
        max_teams:
          type: number
...